refactor: the router is not a target, refactor this into the server directory instead
