+++ integration-tests/aws-sdk/setup.ts
@@ -20,7 +20,7 @@ import { StormDBDataStoreFactory } from "../../src/services/dataStore/stormDb";
 import { otp } from "../../src/services/otp";
 import { JwtTokenGenerator } from "../../src/services/tokenGenerator";
 import { UserPoolServiceFactoryImpl } from "../../src/services/userPoolService";
-import { Router } from "../../src/targets/router";
+import { Router } from "../../src/server/Router";
 
 const mkdtemp = promisify(fs.mkdtemp);
 const rmdir = promisify(fs.rmdir);
