+++ src/targets/adminGetUser.ts
@@ -4,7 +4,7 @@ import {
 } from "aws-sdk/clients/cognitoidentityserviceprovider";
 import { Services } from "../services";
 import { UserNotFoundError } from "../errors";
-import { Target } from "./router";
+import { Target } from "../server/Router";
 
 export type AdminGetUserTarget = Target<
   AdminGetUserRequest,
