+++ src/targets/adminSetUserPassword.ts
@@ -4,7 +4,7 @@ import {
 } from "aws-sdk/clients/cognitoidentityserviceprovider";
 import { UserNotFoundError } from "../errors";
 import { Services } from "../services";
-import { Target } from "./router";
+import { Target } from "../server/Router";
 
 export type AdminSetUserPasswordTarget = Target<
   AdminSetUserPasswordRequest,
