+++ src/targets/deleteUserAttributes.ts
@@ -7,7 +7,7 @@ import { InvalidParameterError, NotAuthorizedError } from "../errors";
 import { Services } from "../services";
 import { Token } from "../services/tokenGenerator";
 import { attributesRemove } from "../services/userPoolService";
-import { Target } from "./router";
+import { Target } from "../server/Router";
 
 export type DeleteUserAttributesTarget = Target<
   DeleteUserAttributesRequest,
