+++ src/targets/getUser.ts
@@ -6,7 +6,7 @@ import jwt from "jsonwebtoken";
 import { InvalidParameterError, UserNotFoundError } from "../errors";
 import { Services } from "../services";
 import { Token } from "../services/tokenGenerator";
-import { Target } from "./router";
+import { Target } from "../server/Router";
 
 export type GetUserTarget = Target<GetUserRequest, GetUserResponse>;
