+++ src/targets/getUserAttributeVerificationCode.ts
@@ -8,7 +8,7 @@ import { InvalidParameterError, UserNotFoundError } from "../errors";
 import { selectAppropriateDeliveryMethod } from "../services/messageDelivery/deliveryMethod";
 import { Token } from "../services/tokenGenerator";
 import { User } from "../services/userPoolService";
-import { Context, Target } from "./router";
+import { Context, Target } from "../server/Router";
 
 const sendAttributeVerificationCode = async (
   ctx: Context,
